+++ Sources/Schemas/OrgComplianceDocumentTypeEnum1.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public enum OrgComplianceDocumentTypeEnum1: String, Codable, Hashable, CaseIterable, Sendable {
+    case securityThreatAssessment = "security_threat_assessment"
+    case insurance
+}