+++ Sources/Schemas/TaskClientCreate1.swift
@@ -4,7 +4,7 @@ public struct TaskClientCreate1: Codable, Hashable, Sendable {
     public let schemaVersion: Int
     public let location: LocationFeature?
     public let action: Action?
-    public let timeWindows: [TimeWindow1]?
+    public let datetimeWindows: [DateTimeWindow1]?
     public let orderPlacerComments: String?
     public let directoryEntryIds: [String]?
     /// Additional properties that are not explicitly defined in the schema
@@ -14,7 +14,7 @@ public struct TaskClientCreate1: Codable, Hashable, Sendable {
         schemaVersion: Int,
         location: LocationFeature? = nil,
         action: Action? = nil,
-        timeWindows: [TimeWindow1]? = nil,
+        datetimeWindows: [DateTimeWindow1]? = nil,
         orderPlacerComments: String? = nil,
         directoryEntryIds: [String]? = nil,
         additionalProperties: [String: JSONValue] = .init()
@@ -22,7 +22,7 @@ public struct TaskClientCreate1: Codable, Hashable, Sendable {
         self.schemaVersion = schemaVersion
         self.location = location
         self.action = action
-        self.timeWindows = timeWindows
+        self.datetimeWindows = datetimeWindows
         self.orderPlacerComments = orderPlacerComments
         self.directoryEntryIds = directoryEntryIds
         self.additionalProperties = additionalProperties
@@ -33,7 +33,7 @@ public struct TaskClientCreate1: Codable, Hashable, Sendable {
         self.schemaVersion = try container.decode(Int.self, forKey: .schemaVersion)
         self.location = try container.decodeIfPresent(LocationFeature.self, forKey: .location)
         self.action = try container.decodeIfPresent(Action.self, forKey: .action)
-        self.timeWindows = try container.decodeIfPresent([TimeWindow1].self, forKey: .timeWindows)
+        self.datetimeWindows = try container.decodeIfPresent([DateTimeWindow1].self, forKey: .datetimeWindows)
         self.orderPlacerComments = try container.decodeIfPresent(String.self, forKey: .orderPlacerComments)
         self.directoryEntryIds = try container.decodeIfPresent([String].self, forKey: .directoryEntryIds)
         self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
@@ -45,7 +45,7 @@ public struct TaskClientCreate1: Codable, Hashable, Sendable {
         try container.encode(self.schemaVersion, forKey: .schemaVersion)
         try container.encodeIfPresent(self.location, forKey: .location)
         try container.encodeIfPresent(self.action, forKey: .action)
-        try container.encodeIfPresent(self.timeWindows, forKey: .timeWindows)
+        try container.encodeIfPresent(self.datetimeWindows, forKey: .datetimeWindows)
         try container.encodeIfPresent(self.orderPlacerComments, forKey: .orderPlacerComments)
         try container.encodeIfPresent(self.directoryEntryIds, forKey: .directoryEntryIds)
     }
@@ -55,7 +55,7 @@ public struct TaskClientCreate1: Codable, Hashable, Sendable {
         case schemaVersion = "schema_version"
         case location
         case action
-        case timeWindows = "time_windows"
+        case datetimeWindows = "datetime_windows"
         case orderPlacerComments = "order_placer_comments"
         case directoryEntryIds = "directory_entry_ids"
     }