+++ Sources/Schemas/TimeWindow1.swift
@@ -1,38 +1,38 @@
 import Foundation
 
 public struct TimeWindow1: Codable, Hashable, Sendable {
-    public let leftBoundaryTimestamp: Date?
-    public let rightBoundaryTimestamp: Date?
+    public let leftBoundaryTime: String?
+    public let rightBoundaryTime: String?
     /// Additional properties that are not explicitly defined in the schema
     public let additionalProperties: [String: JSONValue]
 
     public init(
-        leftBoundaryTimestamp: Date? = nil,
-        rightBoundaryTimestamp: Date? = nil,
+        leftBoundaryTime: String? = nil,
+        rightBoundaryTime: String? = nil,
         additionalProperties: [String: JSONValue] = .init()
     ) {
-        self.leftBoundaryTimestamp = leftBoundaryTimestamp
-        self.rightBoundaryTimestamp = rightBoundaryTimestamp
+        self.leftBoundaryTime = leftBoundaryTime
+        self.rightBoundaryTime = rightBoundaryTime
         self.additionalProperties = additionalProperties
     }
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.leftBoundaryTimestamp = try container.decodeIfPresent(Date.self, forKey: .leftBoundaryTimestamp)
-        self.rightBoundaryTimestamp = try container.decodeIfPresent(Date.self, forKey: .rightBoundaryTimestamp)
+        self.leftBoundaryTime = try container.decodeIfPresent(String.self, forKey: .leftBoundaryTime)
+        self.rightBoundaryTime = try container.decodeIfPresent(String.self, forKey: .rightBoundaryTime)
         self.additionalProperties = try decoder.decodeAdditionalProperties(using: CodingKeys.self)
     }
 
     public func encode(to encoder: Encoder) throws -> Void {
         var container = encoder.container(keyedBy: CodingKeys.self)
         try encoder.encodeAdditionalProperties(self.additionalProperties)
-        try container.encodeIfPresent(self.leftBoundaryTimestamp, forKey: .leftBoundaryTimestamp)
-        try container.encodeIfPresent(self.rightBoundaryTimestamp, forKey: .rightBoundaryTimestamp)
+        try container.encodeIfPresent(self.leftBoundaryTime, forKey: .leftBoundaryTime)
+        try container.encodeIfPresent(self.rightBoundaryTime, forKey: .rightBoundaryTime)
     }
 
     /// Keys for encoding/decoding struct properties.
     enum CodingKeys: String, CodingKey, CaseIterable {
-        case leftBoundaryTimestamp = "left_boundary_timestamp"
-        case rightBoundaryTimestamp = "right_boundary_timestamp"
+        case leftBoundaryTime = "left_boundary_time"
+        case rightBoundaryTime = "right_boundary_time"
     }
 }